* hash.c (rb_any_hash): should treat the return value of rb_objid_hash()   as `long', because ruby assumes the object id of an object is `long'.   this fixes test failures on mswin64 introduced at r44525.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@44534 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ hash.c
@@ -132,7 +132,8 @@ rb_any_hash(VALUE a)
 
     if (SPECIAL_CONST_P(a)) {
 	if (a == Qundef) return 0;
-	hnum = rb_objid_hash((st_index_t)a);
+	/* assume hnum is long, so need to drop upper dword on LLP64. */
+	hnum = (long)rb_objid_hash((st_index_t)a);
     }
     else if (BUILTIN_TYPE(a) == T_STRING) {
 	hnum = rb_str_hash(a);
